Added account setting section, fix route conflicts

+++ database/migrations/2025_05_09_122010_add_user_id_to_dog_profiles_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up()
     {
         Schema::table('dog_profiles', function (Blueprint $table) {
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
+            if (!Schema::hasColumn('dog_profiles', 'user_id')) {
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+            }
         });
     }
